Seperate And Restruct Preview files.
- Preview 를 '텔레그램', '라인', '페이스북'  폴더 별로 나눔
- css 역시 구분화하기 위해 먼저 텔레그램 메인 중 요소들 중 css 적용 되는 className 뒤에 -telegram 을 붙임.
- 따로 Preview.css에서 TelegramPreview.css 를 떼어내어 TelegramPreview 폴더 안에 저장
- 각 테마 별 .js 파일을 만들고 Preview.js에 import 함 (Line, Default, Telegram, Facebook)
- 아직 라인, 페이스북, 기본 ez.ai 인터페이스들 구현 안함
