+++ ez.ai/client/src/components/ChatbotBuild/Preview/DefaultPreview/DefaultPreview.js
@@ -1,0 +1,11 @@
+import React from "react";
+
+const DefaultPreview = () => {
+  return (
+    <>
+    </>
+  )
+
+};
+
+export default DefaultPreview;
