+++ ez.ai/client/src/components/ChatbotBuild/Preview/FacebookPreview/FacebookPreview.js
@@ -1,0 +1,10 @@
+import React from "react";
+
+const FacebookPreview = () => {
+  return (
+    <>
+    </>
+  )
+};
+
+export default FacebookPreview;
