+++ ez.ai/client/src/components/ChatbotBuild/Preview/LinePreview/LinePreview.js
@@ -1,0 +1,9 @@
+import React from "react";
+
+const LinePreview = () => {
+  return (
+    <></>
+  );
+}
+
+export default LinePreview;
