+++ ez.ai/client/src/components/ChatbotBuild/Preview/TelegramPreview/BasicPreview/AudioPreview.js
@@ -12,8 +12,8 @@ const AudioPreview = ({
   return (
       <div className="main-preview">
       <div
-        className={now === i ? "main-content main-audiobox now"
-                            : "main-content main-audiobox"}
+        className={now === i ? "main-content audiobox-telegram now"
+                            : "main-content audiobox-telegram"}
         key={v.content + i}
         onClick={() => {
           setClickedMainInput(v);
@@ -23,12 +23,12 @@ const AudioPreview = ({
       >
         {" "}
         <div>
-          <i className="fas fa-play fa-lg main-file-icon"></i>
-          <div className="main-file-content">
-            <div className="main-file-name" data-filetype="">
+          <i className="fas fa-play fa-lg file-icon-telegram"></i>
+          <div className="file-content-telegram">
+            <div className="file-name" data-filetype="">
               {v.content}
             </div>
-            <div className="main-file-size" data-filetype="">
+            <div className="file-size" data-filetype="">
               {/* 길이, 용량을 표시하려면 데이터를 확장해야함.현재는 파일 이름만 저장 */}
               {/* 00:00, 00.00 MB{" "} */}
             </div>
